Move user routes into router
